pyproject.toml
[tool.poetry]
name = "eva"
version = "0.1.0"
description = ""
authors = ["Paweł Kucmus <pkucmus@gmail.com>"]
packages = [
    { include = "eva_docs", from = "src"},
    { include = "onshape", from = "src"},
]

[tool.poetry.scripts]
eva-docs = 'eva_docs.__main__:cli'

[tool.poetry.plugins."mkdocs.plugins"]
eva-docs-plugin = 'eva_docs.plugin:EVAPlugin'

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
pydantic = "^1.9.0"
cryptography = "^36.0.2"
rich = "^12.2.0"
mkdocs-material = "^8.2.9"
click = "^8.1.2"
sqlmodel = "^0.0.6"
mkdocs-awesome-pages-plugin = "^2.7.0"
mkdocs-minify-plugin = "^0.5.0"
aiofiles = "^0.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipdb = "^0.13.9"
ipython = "^8.2.0"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
